Add Jenkinsfile to build samples, upgrade packages
