+++ tests/testdata/test-applications/coherence/hello-coherence/hello-coherence-app.yaml
