+++ tests/testdata/test-applications/coherence/hello-coherence/openjdk-11.0.2_linux-x64_bin.tar.gz.download
