+++ tests/testdata/test-applications/coherence/hello-coherence/src/main/java/io/verrazzano/coherence/hello/HelloApplication.java
