+++ tests/testdata/test-applications/coherence/hello-coherence/src/main/resources/static/index.html
