--- tests/testdata/test-applications/coherence/hello-verrazzano/src/main/resources/application.properties
@@ -1,4 +1,0 @@
-# Copyright (c) 2022, Oracle and/or its affiliates.
-# Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
-management.endpoints.web.exposure.include=*
-spring.application.name=coherence-springboot-example
