+++ tests/testdata/test-applications/weblogic/hello-weblogic/hello-wls-comp.yaml
