+++ tests/testdata/test-applications/weblogic/hello-weblogic/pom.xml
