+++ tests/testdata/test-applications/weblogic/hello-weblogic/setup/exporter-config.yaml
