+++ tests/testdata/test-applications/weblogic/hello-weblogic/setup/wdt_domain.yaml
