+++ tests/testdata/test-applications/weblogic/hello-weblogic/src/main/java/io/verrazzano/hello/HelloApplication.java
@@ -1,0 +1,21 @@
+// Copyright (c) 2022, Oracle and/or its affiliates.
+// Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
+
+package io.verrazzano.hello;
+
+import io.verrazzano.hello.resource.Greetings;
+
+import java.util.HashSet;
+import java.util.Set;
+import javax.ws.rs.ApplicationPath;
+import javax.ws.rs.core.Application;
+
+@ApplicationPath("/weblogic")
+public class HelloApplication extends Application {
+    public Set<Class<?>> getClasses() {
+        Set<Class<?>> s = new HashSet<>();
+        s.add(Greetings.class);
+        return s;
+    }
+}
+
