+++ tests/testdata/test-applications/weblogic/hello-weblogic/src/main/webapp/index.html
@@ -1,0 +1,11 @@
+<!-- Copyright (c) 2022, Oracle and/or its affiliates. -->
+<!-- Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl. -->
+
+<html>
+  <head>
+    <title>Hello WebLogic Sample</title>
+  </head>
+  <body>
+    Hello WebLogic
+  </body>
+</html>
